+++ R/rec_pattern.R
@@ -27,10 +27,10 @@
 #' # sample data, inspect age of carers
 #' data(efc)
 #' table(efc$c160age, exclude = NULL)
-#' table(rec(efc$c160age, rp$pattern), exclude = NULL)
+#' table(rec(efc$c160age, recodes = rp$pattern), exclude = NULL)
 #'
 #' # recode carers age into groups of width 5
-#' x <- rec(efc$c160age, rp$pattern)
+#' x <- rec(efc$c160age, recodes = rp$pattern)
 #' # add value labels to new vector
 #' set_labels(x) <- rp$labels
 #' # watch result